chore: moves main.py to root folder, gives each class its own file and creates a common folder
