--- src/main.py
@@ -1,265 +1,0 @@
-import os
-import logging
-import argparse
-import re
-import pandas as pd
-from typing import List, Union
-
-logger = logging.getLogger(__name__)
-
-DEFAULT_METHODS = "GET"
-DEFAULT_STATUS = "200,301"
-HTTP_METHODS = ["GET", "POST", "PUT", "DELETE", "HEAD", "OPTIONS", "TRACE"]
-
-def get_formatted_number(num: str) -> int:
-  return "{:,}".format(num)
-
-def get_formatted_time(time_str: str) -> str:
-  base = time_str.split(" ")[0]
-  data = base.split(":")[0]
-  time = ":".join(base.split(":")[1:])
-  return f"{data} {time}"
-
-def get_logs(path: str) -> List[str]:
-  if not os.path.exists(path):
-    raise ValueError(f"File {path} does not exist")
-  
-  with open(path, 'r') as content:
-    return [line.split("\n")[0] for line in content.readlines()]
-
-class LogExtractor:
-  def get_ip(self, log_line: str) -> Union[str, None]:
-    """Extracts the IP address from an Nginx log line."""
-    match = re.match(r'^(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})', log_line)
-    return match.group(1) if match else None
-
-  def get_time(self, log_line: str) -> Union[str, None]:
-    """Extract the date from the Nginx log line"""
-    pattern = r'\[(.*?)\]'
-    matches = re.findall(pattern, log_line)
-
-    if matches is None or len(matches) == 0:
-      return None
-    
-    return get_formatted_time(matches[0])
-
-  def get_method(self, log_line: str) -> Union[str, None]:
-    """Extracts HTTP method from an Nginx log line"""
-    pattern = f'({"|".join(HTTP_METHODS)}) \\/'
-    matches = re.findall(pattern, log_line)
-
-    if matches is None or len(matches) == 0:
-      return None
-    
-    return matches[0].split()[0]
-
-  def get_path(self, log_line: str) -> Union[str, None]:
-    """Extracts the request path/endpoint from an Nginx log line."""
-    whole_path = re.search(r'(?<=] ").+(?=" \d)', log_line) # Match everything between the quotes after the date's `]` and the first digit in the status code
-    
-    if whole_path is None:
-      return None
-    
-    whole_path = whole_path.group(0)
-
-    # remove method if it exists
-    if whole_path.upper().startswith(tuple(HTTP_METHODS)):
-      items = whole_path.split(" ")[1:]
-      whole_path = " ".join(items)
-
-    # remove the protocol if it exists
-    protocol = re.search(r'HTTP\/\d+\.\d+', whole_path)
-    if protocol is not None:
-      items = whole_path.split(" ")
-      items.pop()
-      whole_path = " ".join(items)
-
-    return whole_path.strip()
-
-  def get_status_code(self, log_line: str) -> Union[str, None]:
-    """Extracts the HTTP status code from an Nginx log line."""
-    match = re.search(r'(?<=(HTTP\/1.0|HTTP\/1.1|HTTP\/2.0)" )[0-9]+', log_line) # Match 3 digit code after HTTP protocol
-
-    if match is None:
-      match = re.search(r'(?<=" )[0-9]+', log_line) # Match 3 digit code after HTTP protocol quotation
-
-    return match[0] if match else None
-
-  def get_size(self, log_line: str) -> Union[str, None]:
-    """Extracts the response size in bytes from an Nginx log"""
-    match = re.search(r'/(?<=(HTTP\/1.0|HTTP\/1.1|HTTP\/2.0)" )[0-9]+ [0-9]+', log_line) # Match multi two sets of digits after HTTP protocol
-
-    if match is None:
-      match = re.search(r'(?<=" )[0-9]+ [0-9]+', log_line) # Match multi two sets of digits after HTTP protocol quotation
-
-    return match[0].split(" ")[1] if match else None
-
-  def get_agent(self, log_line: str) -> Union[str, None]:
-    """Extracts the user agent string and attempts to identify the device from it."""
-    match = re.search(r'(?<=" ").+(?="$)', log_line)
-    user_agent = match[0] if match else None
-    if user_agent:
-      if "Android" in user_agent:
-        return "Android"
-      elif "iPhone" in user_agent or "iPad" in user_agent:
-        return "iOS"
-      elif "Windows" in user_agent:
-        return "Windows"
-      elif "Mac" in user_agent:
-        return "Mac"
-      elif "-" == user_agent:
-        return "N/A"
-      else:
-        match = re.search(r'(?<=\()[a-zA-Z0-9\s\.:;\-\/_]+(?=\))', user_agent)
-        return match[0] if match else "Unknown"
-    return None
-
-  def to_object(self, log_line: str) -> dict:
-    """Returns a dictionary containing the log data."""
-    
-    return {
-      "ip": self.get_ip(log_line),
-      "time": self.get_time(log_line),
-      "method": self.get_method(log_line),
-      "path": self.get_path(log_line),
-      "status_code": self.get_status_code(log_line),
-      "response_size": self.get_size(log_line),
-      "user_agent": self.get_agent(log_line)
-      }
-
-class LogAnalyzer:
-  extractor = LogExtractor()
-
-  def __init__(self, logs: List[str], verbose: bool):
-    self.logs = logs
-    self.verbose = verbose
-
-  def find_logs_with_approved_methods(self, approved_methods: List[str], inverse: bool = False) -> List[str]:
-    matching_logs = []
-    for log in self.logs:
-      current_method = self.extractor.get_method(log)
-      if inverse and current_method not in approved_methods:
-        matching_logs.append(log)
-      elif inverse is False and current_method in approved_methods:
-        matching_logs.append(log)
-    return matching_logs
-    
-  def find_logs_with_approved_status(self, approved_status_codes: List[int], inverse: bool = False) -> List[str]:
-    matching_logs = []
-    for log in self.logs:
-      current_status = self.extractor.get_status_code(log)
-      if inverse and current_status not in approved_status_codes:
-        matching_logs.append(log)
-      elif inverse is False and current_status in approved_status_codes:
-        matching_logs.append(log)
-    return matching_logs
-
-  def find_sus_paths(self) -> List[str]:
-    matching_logs = []
-    for log in self.logs:
-      current_path = self.extractor.get_path(log)
-      if current_path is None or current_path.strip() == "":
-        matching_logs.append(log) # if you cant find a valid path, its sus
-        continue
-      sus_file = re.findall(r'(?<=\/)[\w|\d]*\.[\w|\d]*', current_path)
-      if sus_file is None or len(sus_file) == 0:
-        continue # sus file not found, skip to next log
-      sus_file = str(sus_file[0].lower().strip())
-      if sus_file.endswith(".html"):
-        continue # this is most likely a normal HTML file
-      elif sus_file.endswith((".jpg", ".jpeg", ".png", ".svg", ".webp", ".ico", ".gif", ".mp3", ".mp4", ".mov", ".avi", ".pdf", ".doc", ".docx", ".xls", ".xlsx", ".zip")):
-        continue # this is most likely a normal media file
-      elif current_path.startswith(("/_nuxt/", "/api/_nuxt/", "/api/_nuxt_icon")) and sus_file.endswith((".js", ".json", ".css")):
-        continue # this is most likely a bundled js file
-      elif sus_file.endswith(".php") and sus_file not in ["config.php", "db.php", "auth.php", "shell.php", "cmd.php", "backup.php", "connect.php", "index_backup.php", "main.php", "upload.php", "reset.php", "install.php"]:
-        continue # this is most likely a normal php file
-      elif sus_file.find("\\") != -1:
-        matching_logs.append(log) # a backwards slash is super sus, could be obfuscated malicious commands
-      else:
-        matching_logs.append(log)
-    return matching_logs
-
-def report(title: str, description: str, logs: List[str], max_logs: int = 10, format: str = "table", verbose: bool = False) -> str:
-  extractor = LogExtractor()
-  logs = [extractor.to_object(log) for log in logs] # convert logs to objects
-
-  df = pd.DataFrame(logs) if verbose else pd.DataFrame(logs[:max_logs])
-
-  if format == "csv":
-    # remove index column
-    df = df.to_csv(index=False)
-
-  report = f"Title: {title}\nDescription: {description}\n\n{df}"
-    
-  if len(logs) > max_logs and verbose is False:
-    report += f"\n\n... ({get_formatted_number(len(logs) - max_logs)} more)"
-
-
-  if len(logs) > 0:
-    report += f"\n\n[count: {get_formatted_number(len(logs))}]"
-  else:
-    report += "No logs found"
-
-  return report
-
-if __name__ == "__main__":
-    logging.basicConfig(level=logging.INFO)
-    
-    parser = argparse.ArgumentParser(prog="Basset",description="Analyze your Nginx logs")
-    parser.add_argument("-a", "--analysis", default="all", choices=['all', 'methods', 'status', 'paths'], help="The type of analysis to perform (defaults to 'all')")
-    parser.add_argument("-f", "--file", required=True, help="Path to the log file")
-    parser.add_argument("-s", "--status", default=DEFAULT_STATUS, help="Specify normal HTTP status codes comma-separated")
-    parser.add_argument("-m", "--methods", default=DEFAULT_METHODS, help="Specify normal HTTP methods comma-separated")
-    parser.add_argument("-F", "--format", default="table", choices=['table', 'csv'], help="The type of format to dispaly the data"),
-    parser.add_argument("-v", "--verbose", action="store_true", help="Show extensive reports")
-    parser.add_argument("-d", "--debug", action="store_true", help="Show debug logs")
-
-    args = parser.parse_args()
-    
-    if args.debug:
-      logger.setLevel(logging.DEBUG)
-      for key in args.__dict__.keys():
-        logger.debug(f"[config] {key}: {args.__dict__[key]}")
-
-    try:
-      analyzer = LogAnalyzer(logs=get_logs(args.file), verbose=args.verbose)
-      reports = []
-
-      if args.analysis == "all" or args.analysis == "methods":
-        methods = args.methods.split(",")
-        reports.append(report(
-            title="Suspicious Methods",
-            description=f"Logs with HTTP methods that are not {','.join(methods)}",
-            logs=analyzer.find_logs_with_approved_methods(approved_methods=methods, inverse=True),
-            format=args.format,
-            verbose=args.verbose
-        ))
-
-      if args.analysis == "all" or args.analysis == "status":
-        status_codes = args.status.split(",")
-        reports.append(report(
-            title="Suspicious Status",
-            description=f"Logs with HTTP status codes that are not {','.join(status_codes)}",
-            logs=analyzer.find_logs_with_approved_status(approved_status_codes=status_codes, inverse=True),
-            format=args.format,
-            verbose=args.verbose
-        ))
-
-      if args.analysis == "all" or args.analysis == "paths":
-        reports.append(report(
-            title="Suspicious Paths",
-            description=f"Logs with suspicious paths",
-            logs=analyzer.find_sus_paths(),
-            format=args.format,
-            verbose=args.verbose
-        ))
-
-      for report in reports:
-          print("\n==================")
-          print(report)
-      
-      print(f"\n\nTotal logs: {get_formatted_number(len(analyzer.logs))}")
-    except ValueError as e:
-        logger.error(f"error: {e}")
-        parser.print_help()
-        parser.exit(status=1)
